requirements.txt
# Utilities
PyYAML==6.0.1
requests==2.31.0
watchdog==3.0.0
APScheduler==3.10.4

# Web Server
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# OCR Tools
langdetect==1.0.9
pytesseract==0.3.10
pdf2image==1.17.0

# Text-Verarbeitung
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# Data Analysis & Export
pandas==2.1.4
xlsxwriter==3.1.9
reportlab==4.0.8

# Datum-Extraktion
dateparser==1.2.0
regex==2023.10.3

# Logging
colorlog==6.8.0

# System Monitor
psutil==5.9.6

# Image Processing (OCR Pre-Processing)
opencv-python==4.8.1.78

# Metrics & Monitoring
prometheus-client==0.19.0

# Advanced OCR
easyocr==1.7.1

# Semantic Search
sentence-transformers>=3.0.0
numpy==1.26.0
scikit-learn==1.3.2

# ORM & Database Migrations
SQLAlchemy==2.0.23
alembic==1.13.1
